Refactor token auth to source-specific classes
